Move StrictWarnings to the tools directory
Followup: https://github.com/rails/rails/pull/46189

Not sure why I put it in Active Support lib directory,
that makes little sense, as it's shipped with the gem
but isn't something usable outside of rails itself.
