+++ actionmailer/test/abstract_unit.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "active_support/testing/strict_warnings"
+require_relative "../../tools/strict_warnings"
 require "active_support/core_ext/kernel/reporting"
 
 # These are the normal settings that will be set up by Railties
