+++ railties/test/generators/generator_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "active_support/testing/strict_warnings"
+require_relative "../../../tools/strict_warnings"
 require "active_support/test_case"
 require "active_support/testing/autorun"
 require "rails/generators/app_base"
